# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_emzola_issuetracker_internal_model.Project:
    properties:
      actual_end_date:
        type: string
      assigned_to:
        type: integer
      created_by:
        type: string
      created_on:
        type: string
      description:
        type: string
      id:
        type: integer
      modified_by:
        type: string
      modified_on:
        type: string
      name:
        type: string
      start_date:
        type: string
      target_end_date:
        type: string
    type: object
  github_com_emzola_issuetracker_internal_model.Token:
    properties:
      token:
        type: string
    type: object
  github_com_emzola_issuetracker_internal_model.User:
    properties:
      activated:
        type: boolean
      created_by:
        type: string
      created_on:
        type: string
      email:
        type: string
      id:
        type: integer
      modified_by:
        type: string
      modified_on:
        type: string
      name:
        type: string
      role:
        type: string
    type: object
...
  internal_handler_http.createActivationTokenPayload:
    properties:
      email:
        type: string
    type: object
  internal_handler_http.createAuthenticationTokenPayload:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  internal_handler_http.createProjectPayload:
    properties:
      assigned_to:
        type: integer
      description:
        type: string
      name:
        type: string
      start_date:
        type: string
      target_end_date:
        type: string
    type: object
...
  internal_handler_http.updateProjectPayload:
    properties:
      actual_end_date:
        type: string
      assigned_to:
        type: integer
      description:
        type: string
      name:
        type: string
      start_date:
        type: string
      target_end_date:
        type: string
    type: object
...
host: https://issuetracker-api-dev.fl0.io
info:
  contact:
    email: emma.idika@yahoo.com
    name: API Support
  description: This is an API service for an issue tracker.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Issue Tracker API
  version: 1.0.0
paths:
...
  /v1/projects:
    get:
      description: This endpoint gets all projects
      parameters:
      - description: Bearer token
        in: header
        name: token
        required: true
        type: string
      - description: Query string param for name
        in: query
        name: name
        type: string
      - description: Query string param for assigned_to
        in: query
        name: assigned_to
        type: string
      - description: Query string param for start_Date
        in: query
        name: start_date
        type: string
      - description: Query string param for target_end_date
        in: query
        name: target_end_date
        type: string
      - description: Query string param for actual_end_date
        in: query
        name: actual_end_date
        type: string
      - description: Query string param for created_by
        in: query
        name: created_by
        type: string
      - description: Query string param for pagination (min 1)
        in: query
        name: page
        type: string
      - description: Query string param for pagination (max 100)
        in: query
        name: page_size
        type: string
      - description: 'Sort by asc or desc order. Asc: id, name, assigned_to, start_date,
          target_end_date, actual_end_date, created_by | Desc: -id, -name, -assigned_to,
          -start_date, -target_end_date, -actual_end_date, -created_by'
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_emzola_issuetracker_internal_model.Project'
            type: array
        "422":
          description: Unprocessable Entity
        "500":
          description: Internal Server Error
      summary: Get all projects
      tags:
      - projects
    post:
      consumes:
      - application/json
      description: Create a new project with the request payload
      parameters:
      - description: Bearer token
        in: header
        name: token
        required: true
        type: string
      - description: Request payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_handler_http.createProjectPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_emzola_issuetracker_internal_model.Project'
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "422":
          description: Unprocessable Entity
        "500":
          description: Internal Server Error
      summary: Create a new project
      tags:
      - projects
  /v1/projects/{project_id}:
    delete:
      description: This endpoint deletes a project
      parameters:
      - description: Bearer token
        in: header
        name: token
        required: true
        type: string
      - description: ID of project to delete
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Delete a project
      tags:
      - projects
    get:
      description: This endpoint gets a project by ID
      parameters:
      - description: Bearer token
        in: header
        name: token
        required: true
        type: string
      - description: ID of project to get
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_emzola_issuetracker_internal_model.Project'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get project by ID
      tags:
      - projects
    patch:
      consumes:
      - application/json
      description: This endpoint updates a project
      parameters:
      - description: Bearer token
        in: header
        name: token
        required: true
        type: string
      - description: Request payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_handler_http.updateProjectPayload'
      - description: ID of project to update
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_emzola_issuetracker_internal_model.Project'
        "400":
          description: Bad Request
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "409":
          description: Conflict
        "422":
          description: Unprocessable Entity
        "500":
          description: Internal Server Error
      summary: Update a project
      tags:
      - projects
  /v1/projects/{project_id}/users:
    get:
      description: This endpoint gets all project users
      parameters:
      - description: Bearer token
        in: header
        name: token
        required: true
        type: string
      - description: ID of project to get users
        in: path
        name: project_id
        required: true
        type: string
      - description: Query string param for role
        in: query
        name: role
        type: string
      - description: Query string param for pagination (min 1)
        in: query
        name: page
        type: string
      - description: Query string param for pagination (max 100)
        in: query
        name: page_size
        type: string
      - description: 'Sort by asc or desc order. Asc: id | Desc: -id'
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_emzola_issuetracker_internal_model.User'
            type: array
        "422":
          description: Unprocessable Entity
        "500":
          description: Internal Server Error
      summary: Get project users
      tags:
      - projects
  /v1/tokens/activation:
    post:
      consumes:
      - application/json
      description: This endpoint creates a new activation token
      parameters:
      - description: Request payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_handler_http.createActivationTokenPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "422":
          description: Unprocessable Entity
        "500":
          description: Internal Server Error
      summary: Create a new activation token
      tags:
      - tokens
  /v1/tokens/authentication:
    post:
      consumes:
      - application/json
      description: This endpoint creates JWT token
      parameters:
      - description: Request payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_handler_http.createAuthenticationTokenPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_emzola_issuetracker_internal_model.Token'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "422":
          description: Unprocessable Entity
        "500":
          description: Internal Server Error
      summary: Create JWT authentication token
      tags:
      - tokens
...
swagger: "2.0"